RemoteHostをモジュールに分離
- remote_hostモジュールを新設。RemoteHost関係をすべて移設。
- URI形式のサポートに備え、RemoteHostにportを追加。
- portが二箇所に分離したことに対するssh_connect::get_addressの調整
