+++ Cargo.toml
@@ -19,6 +19,7 @@ daemonize = "0.5.0"
 dialoguer = "0.12.0"
 dns-lookup = "3.0.0"
 env_logger = "0.11.0"
+fluent-uri = "0.4.1"
 fuser = "0.16"
 home = "0.5.4"
 libc = "0.2.139"
